Split environment-independent logics into Player
