+++ lib/glsl-livecoder.js
@@ -3,11 +3,11 @@ import path from 'path';
 import glslify from 'glslify';
 import GlslLivecoderView from './glsl-livecoder-view';
 import { validator, loadFile } from './validator';
-import ThreeShader from './three-shader';
 import Config from './config';
 import type { Pass } from './three-shader';
 import type { RcDiff } from './config';
-import { INITIAL_FRAGMENT_SHADER, DEFAULT_VERTEX_SHADER } from './constants';
+import { INITIAL_FRAGMENT_SHADER } from './constants';
+import Player from './player';
 
 declare var atom: any;
 type TextEditor = any;
@@ -19,8 +19,7 @@ type GlslLivecoderState = {
 };
 
 export default class GlslLivecoder {
-  _view: GlslLivecoderView;
-  _three: ThreeShader;
+  _player: Player;
   _state: GlslLivecoderState;
   _glslangValidatorPath: string;
   _lastShader: Pass[];
@@ -28,8 +27,10 @@ export default class GlslLivecoder {
   _config: Config;
 
   constructor(config: Config) {
-    this._view = new GlslLivecoderView();
-    atom.workspace.element.appendChild(this._view.getElement());
+    const view = new GlslLivecoderView();
+    atom.workspace.element.appendChild(view.getElement());
+
+    this._player = new Player(view, config);
 
     this._config = config;
     this._config.on('change', this._onChange);
@@ -37,86 +38,31 @@ export default class GlslLivecoder {
     const rc = this._config.rc;
     this._glslangValidatorPath = rc.glslangValidatorPath;
 
-    this._three = new ThreeShader(rc);
-    this._three.setCanvas(this._view.getCanvas());
-
     this._state = {
       isPlaying: false,
     };
+
+    this._player.loadShader([{
+      fs: INITIAL_FRAGMENT_SHADER,
+    }]);
+    this.play();
   }
 
   destroy(): void {
-    this._three.stop();
-    this._view.destroy();
+    this._player.destroy();
   }
 
   _onChange = ({ newConfig, added, removed }: RcDiff) => {
-    console.log('Update config', newConfig);
-    // Get paths for videos still in use
-    const importedPaths = {};
-    Object.values(newConfig.IMPORTED).forEach(imported => {
-      importedPaths[(imported: any).PATH] = true;
-    });
-
-    Object.keys(removed.IMPORTED).forEach(key => {
-      const path = removed.IMPORTED[key].PATH;
-      this._three.unloadTexture(key, path, !importedPaths[path]);
-    });
-    Object.keys(added.IMPORTED || {}).forEach(key => {
-      this._three.loadTexture(key, added.IMPORTED[key].PATH);
-    });
-    if (added.vertexMode) {
-      this._three.setVertexMode(added.vertexMode);
-    }
-    if (added.vertexCount) {
-      this._three.setVertexCount(added.vertexCount);
-    }
-    if (added.pixelRatio) {
-      this._three.setPixelRatio(added.pixelRatio);
-    }
-    if (added.frameskip) {
-      this._three.setFrameskip(added.frameskip);
-    }
     if (added.glslangValidatorPath) {
       this._glslangValidatorPath = added.glslangValidatorPath;
     }
-    if (added.audio !== undefined) {
-      this._three.toggleAudio(added.audio);
-    }
-    if (added.midi !== undefined) {
-      this._three.toggleMidi(added.midi);
-    }
-    if (added.keyboard !== undefined) {
-      this._three.toggleKeyboard(added.keyboard);
-    }
-    if (added.gamepad !== undefined) {
-      this._three.toggleGamepad(added.gamepad);
-    }
-    if (added.camera !== undefined) {
-      this._three.toggleCamera(added.camera);
-    }
 
+    this._player.onChange({ newConfig, added, removed });
     this._loadLastShader();
   }
 
   setValidatorPath(path: string): void {
     this._glslangValidatorPath = path;
-  }
-
-  setPixelRatio(x: number): void {
-    this._three.setPixelRatio(x);
-  }
-
-  setFrameskip(x: number): void {
-    this._three.setFrameskip(x);
-  }
-
-  setVertexCount(x: number): void {
-    this._three.setVertexCount(x);
-  }
-
-  setVertexMode(x: string): void {
-    this._three.setVertexMode(x);
   }
 
   toggle(): void {
@@ -128,19 +74,13 @@ export default class GlslLivecoder {
   }
 
   play(): void {
-    this._view.show();
-    this._three.loadShader([{
-      vs: DEFAULT_VERTEX_SHADER,
-      fs: INITIAL_FRAGMENT_SHADER,
-    }]);
-    this._three.play();
     this._state.isPlaying = true;
+    this._player.play();
   }
 
   stop(): void {
-    this._view.hide();
-    this._three.stop();
     this._state.isPlaying = false;
+    this._player.stop();
     this.stopWatching();
     this._config.reset();
   }
@@ -182,7 +122,7 @@ export default class GlslLivecoder {
     if (!this._lastShader) {
       return;
     }
-    this._three.loadShader(this._lastShader);
+    this._player.loadShader(this._lastShader);
   }
 
   stopWatching(): void {
@@ -261,7 +201,7 @@ export default class GlslLivecoder {
         }));
       })
       .then(passes => {
-        this._three.loadShader(passes);
+        this._player.loadShader(passes);
         this._lastShader = passes;
       })
       .catch(e => {
